群组跟环信对接

+++ src/main/java/com/award/sy/web/openapi/ChatOpenController.java
@@ -13,10 +13,12 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
+import com.award.core.util.ImUtils;
 import com.award.core.util.JsonUtils;
 import com.award.sy.common.DateUtil;
 import com.award.sy.entity.Group;
 import com.award.sy.entity.GroupDetails;
+import com.award.sy.entity.User;
 import com.award.sy.service.FriendService;
 import com.award.sy.service.GroupDetailsService;
 import com.award.sy.service.GroupService;
@@ -57,6 +59,7 @@ public class ChatOpenController {
 		boolean isFriend = true;
 		if(!StringUtils.isBlank(userId) && !StringUtils.isBlank(members) && !StringUtils.isBlank(groupName)) {
 			String[] member = members.split(",");
+			if(member.length >= 2) {
 			for(int i = 0; i < member.length; i++) {
 				List<Map<String,Object>> list = friendService.getFriendByTwoId(Long.parseLong(userId), Long.parseLong(member[i]));
 				if(list.size() > 0) {
@@ -69,11 +72,45 @@ public class ChatOpenController {
 				}
 			}
 			if(isFriend) {
-				Group group = new Group();
+				String[] ids = userId.concat(",").concat(members).split(",");
+				List<User> userList = userService.getUserByIds(ids);
+				String memberNames = "";
+				String admin = "";
+				for(User user:userList) {
+					if(user.getUser_id() == Long.parseLong(userId)) {
+						admin = user.getUser_name();
+					}
+					String[] memberStr = members.split(",");
+					for(int i = 0; i < memberStr.length; i++) {
+						if(user.getUser_id() == Long.parseLong(memberStr[i])) {
+							if(memberNames.equals("")) {
+								memberNames = user.getUser_name();
+							}else {
+								memberNames = memberNames.concat(",").concat(user.getUser_name());
+							}
+						}
+					}
+				}
+				Group group = groupService.getGroupByNameAndAdminId(groupName,Long.parseLong(userId));
+				if(group != null) {
+					return JsonUtils.writeJson(0,33, "该群已存在");
+				}
+				String imGroupId = "";
+				if(!admin.equals("") && !memberNames.equals("")) {
+					imGroupId = ImUtils.addGroup(groupName, memberNames.split(","), admin);
+					if(imGroupId.equals("")) {
+						return JsonUtils.writeJson(0,27, "环信群创建失败");
+					}
+				}else {
+					return JsonUtils.writeJson(0,26, "群成员不存在");
+				}				
+				group = new Group();
 				group.setGroup_name(groupName);
 				group.setCreate_time(DateUtil.getNowTime());
+				group.setIm_group_id(imGroupId);
+				group.setAdmin_id(Long.parseLong(userId));
 				groupService.addGroup(group);
-				Group group2 = groupService.getGroupByName(groupName);
+				Group group2 = groupService.getGroupByNameAndAdminId(groupName,Long.parseLong(userId));
 				GroupDetails gd = new GroupDetails();
 				gd.setMember_id(Long.parseLong(userId));
 				gd.setIs_admin(1);
@@ -85,13 +122,16 @@ public class ChatOpenController {
 					gd2.setMember_id(Long.parseLong(member[i]));
 					gd2.setIs_admin(0);
 					gd2.setJoin_time(DateUtil.getNowTime());
+					gd2.setGroup_id(group2.getGroup_id());
 					groupDetailsService.addGroupDetails(gd2);
 				}
 				returnStr = JsonUtils.writeJson("创建成功",1);
 			}else {
 				returnStr = JsonUtils.writeJson(0,7, "存在与你不是互关的成员");
 			}
-			
+			}else {
+				returnStr = JsonUtils.writeJson(0,32, "群组成员不能小于3个");
+			}
 		}
 		return returnStr;
 	}
@@ -104,6 +144,8 @@ public class ChatOpenController {
 		String groupId = request.getParameter("groupId");
 		String returnStr = JsonUtils.writeJson(0, 0, "参数为空");
 		if(!StringUtils.isBlank(userId) && !StringUtils.isBlank(member) && !StringUtils.isBlank(groupId)) {
+			Group group = groupService.getGroupById(Long.parseLong(groupId));
+			if(group != null) {
 			GroupDetails gd = groupDetailsService.getGroupDetailsByGroupIdAndUserId(Long.parseLong(groupId), Long.parseLong(userId));
 			if(gd != null) {
 				if(gd.getIs_admin() == 0) {
@@ -111,8 +153,18 @@ public class ChatOpenController {
 				}else {
 					GroupDetails gd1 = groupDetailsService.getGroupDetailsByGroupIdAndUserId(Long.parseLong(groupId), Long.parseLong(member));
 					if(gd1 != null) {
-						groupDetailsService.removeGroupDetails(gd1.getDetails_id());
-						returnStr = JsonUtils.writeJson("删除成功",1);
+						User user = userService.getUserById(Long.parseLong(member));
+						if(user != null) {
+						boolean isSuccess = ImUtils.delSingleMember(group.getIm_group_id(), user.getUser_name());
+						if(isSuccess) {
+							groupDetailsService.removeGroupDetails(gd1.getDetails_id());
+						    returnStr = JsonUtils.writeJson("删除成功",1);
+						}else {
+							returnStr = JsonUtils.writeJson(0, 29, "环信群成员删除失败");
+						}
+						}else {
+							returnStr = JsonUtils.writeJson(0, 26, "成员账号不存在");
+						}
 					}else {
 						returnStr = JsonUtils.writeJson(0, 15, "成员ID不存在");
 					}
@@ -120,6 +172,53 @@ public class ChatOpenController {
 			}else {
 				returnStr = JsonUtils.writeJson(0, 13, "群主ID不存在");
 			}
+		}else {
+			returnStr = JsonUtils.writeJson(0, 25, "群ID不存在");
+		}
+		}
+		return returnStr;
+	}
+	
+	@RequestMapping(value = "/open/addMember", produces = "text/html;charset=UTF-8")
+	@ResponseBody
+	public String addMember(HttpServletRequest request) {
+		String member = request.getParameter("member");
+		String userId = request.getParameter("userId");
+		String groupId = request.getParameter("groupId");
+		String returnStr = JsonUtils.writeJson(0, 0, "参数为空");
+		if(!StringUtils.isBlank(userId) && !StringUtils.isBlank(member) && !StringUtils.isBlank(groupId)) {
+			Group group = groupService.getGroupById(Long.parseLong(groupId));
+			if(group != null) {
+			GroupDetails gd = groupDetailsService.getGroupDetailsByGroupIdAndUserId(Long.parseLong(groupId), Long.parseLong(userId));
+			if(gd != null) {
+					GroupDetails gd1 = groupDetailsService.getGroupDetailsByGroupIdAndUserId(Long.parseLong(groupId), Long.parseLong(member));
+					if(gd1 == null) {
+						User user = userService.getUserById(Long.parseLong(member));
+						if(user != null) {
+						boolean isSuccess = ImUtils.addSingleMember(group.getIm_group_id(), user.getUser_name());
+						if(isSuccess) {
+							gd1 = new GroupDetails();
+							gd1.setGroup_id(Long.parseLong(groupId));
+							gd1.setIs_admin(0);
+							gd1.setJoin_time(DateUtil.getNowTime());
+							gd1.setMember_id(Long.parseLong(member));
+							groupDetailsService.addGroupDetails(gd1);
+						    returnStr = JsonUtils.writeJson("添加成功",1);
+						}else {
+							returnStr = JsonUtils.writeJson(0, 31, "环信群成员添加失败");
+						}
+						}else {
+							returnStr = JsonUtils.writeJson(0, 26, "成员账号不存在");
+						}
+					}else {
+						returnStr = JsonUtils.writeJson(0, 30, "成员已经存在");
+					}
+			}else {
+				returnStr = JsonUtils.writeJson(0, 15, "邀请人不存在");
+			}
+		}else {
+			returnStr = JsonUtils.writeJson(0, 25, "群ID不存在");
+		}
 		}
 		return returnStr;
 	}
@@ -137,6 +236,10 @@ public class ChatOpenController {
 				GroupDetails gd = groupDetailsService.getGroupDetailsByGroupIdAndUserId(Long.parseLong(groupId), Long.parseLong(userId));
 				if(gd != null) {
 					if(gd.getIs_admin() == 1) {
+						boolean isSuccess = ImUtils.updateGroup(groupName, group.getIm_group_id());
+						if(!isSuccess) {
+							return JsonUtils.writeJson(0, 28, "修改失败");
+						}
 						group.setGroup_name(groupName);
 						groupService.editGroup(group);
 						returnStr = JsonUtils.writeJson("修改成功",1);
